refactor AndroidChatService, merge packages
