+++ qabelbox/src/main/java/de/qabel/qabelbox/chat/services/AndroidChatServiceResponder.kt
@@ -1,4 +1,4 @@
-package de.qabel.qabelbox.chat.service
+package de.qabel.qabelbox.chat.services
 
 import android.content.BroadcastReceiver
 import android.content.Context
@@ -7,8 +7,11 @@ import de.qabel.qabelbox.QblBroadcastConstants
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
 
-
-class ChatServiceStarter : BroadcastReceiver(), AnkoLogger {
+/**
+ * Starts the Service on receiving QblBroadcastConstants.Chat.INTENT_SHOW_NOTIFICATION
+ * to show Notifications.
+ */
+class AndroidChatServiceResponder : BroadcastReceiver(), AnkoLogger {
 
     override fun onReceive(context: Context, intent: Intent?) {
         info("Receive notify broadcast. Start service")
