+++ qabelbox/src/main/java/de/qabel/qabelbox/chat/view/views/ChatFragment.kt
@@ -58,7 +58,7 @@ class ChatFragment : ChatView, BaseFragment(), AnkoLogger {
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        contactKeyId = arguments.getString(ARG_CONTACT)?: throw IllegalArgumentException(
+        contactKeyId = arguments.getString(ARG_CONTACT) ?: throw IllegalArgumentException(
                 "Starting ChatFragment without contactKeyId")
         val component = getComponent(MainActivityComponent::class.java).plus(ChatModule(this))
         component.inject(this)
@@ -85,14 +85,18 @@ class ChatFragment : ChatView, BaseFragment(), AnkoLogger {
      * Block notifications in which only the currently active contact
      * and identity are involved.
      */
-    private val notificationBlockReceiver = object: BroadcastReceiver() {
-        override fun onReceive(context: Context?, intent: Intent?) {
+    private val notificationBlockReceiver = object : BroadcastReceiver() {
+        override fun onReceive(context: Context, intent: Intent) {
             if (isOrderedBroadcast) {
-                val ids = intent?.getStringArrayListExtra(Helper.AFFECTED_IDENTITIES_AND_CONTACTS)
+                val ids = intent.getStringArrayListExtra(Helper.AFFECTED_IDENTITIES_AND_CONTACTS)
                         ?.filterNotNull() ?: return
-                if (ids.any({(it == contactKeyId) or (it == identity.keyIdentifier)})) {
-                    abortBroadcast()
-                    if(injectCompleted){
+
+                val currentKeys = listOf(contactKeyId, identity.keyIdentifier)
+                if (ids.containsAll(currentKeys)) {
+                    if (ids.all { currentKeys.contains(it) }) {
+                        abortBroadcast()
+                    }
+                    if (injectCompleted) {
                         presenter.refreshMessages()
                     }
                 }
@@ -153,7 +157,11 @@ class ChatFragment : ChatView, BaseFragment(), AnkoLogger {
         adapter.notifyDataSetChanged()
     }
 
-    override val title: String by lazy {if (injectCompleted) presenter.title else { "" } }
+    override val title: String by lazy {
+        if (injectCompleted) presenter.title else {
+            ""
+        }
+    }
 
     override fun supportBackButton(): Boolean = true
