+++ qabelbox/src/main/java/de/qabel/qabelbox/dagger/components/ApplicationComponent.java
@@ -10,7 +10,7 @@ import de.qabel.core.repository.IdentityRepository;
 import de.qabel.qabelbox.activities.CreateAccountActivity;
 import de.qabel.qabelbox.activities.CreateIdentityActivity;
 import de.qabel.qabelbox.activities.SplashActivity;
-import de.qabel.qabelbox.chat.service.AndroidChatService;
+import de.qabel.qabelbox.chat.services.AndroidChatService;
 import de.qabel.qabelbox.chat.services.QabelFirebaseMessagingService;
 import de.qabel.qabelbox.sync.QabelSyncAdapter;
 import de.qabel.qabelbox.dagger.modules.AccountModule;
